Document `textx_isinstance` and move to model module
