+++ textx/__init__.py
@@ -1,10 +1,9 @@
 # flake8: noqa
 from textx.metamodel import metamodel_from_file, metamodel_from_str
 from textx.model import get_children_of_type, get_parent_of_type, \
-    get_model, get_metamodel, get_children, get_location
+    get_model, get_metamodel, get_children, get_location, textx_isinstance
 from textx.exceptions import TextXError, TextXSyntaxError, \
     TextXSemanticError, TextXRegistrationError
-from textx.scoping.tools import textx_isinstance
 from textx.generators import get_output_filename, gen_file
 from textx.registration import (LanguageDesc, GeneratorDesc,
                                 register_language, register_generator,
